+++ packages/x-utils/src/object.ts
@@ -260,3 +260,38 @@ interface RenameOptions<Prefix, Suffix> {
   prefix?: Prefix;
   suffix?: Suffix;
 }
+
+/**
+ * Checks if two objects are deeply equal.
+ *
+ * @param object1 - First object to compare.
+ * @param object2 - Second object to compare.
+ *
+ * @returns True if both objects are deeply equal. False otherwise.
+ * @public
+ */
+export function deepEqual<ObjectType extends Dictionary>(
+  object1: ObjectType | undefined,
+  object2: ObjectType | undefined
+): boolean {
+  if (object1 === object2) {
+    return true;
+  }
+
+  if (!object1 || !object2 || typeof object1 !== 'object' || typeof object2 !== 'object') {
+    return false;
+  }
+
+  const keys1 = Object.keys(object1);
+  const keys2 = Object.keys(object2);
+
+  if (keys1.length !== keys2.length) {
+    return false;
+  }
+
+  return keys1.length !== keys2.length
+    ? false
+    : keys1.every(key => {
+        return keys2.includes(key) && deepEqual(object1[key], object2[key]);
+      });
+}