+++ packages/deep-merge/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.0-alpha.3](https://github.com/empathyco/x/compare/@empathyco/x-deep-merge@2.0.0-alpha.2...@empathyco/x-deep-merge@2.0.0-alpha.3) (2023-09-18)
+
+**Note:** Version bump only for package @empathyco/x-deep-merge
+
+
+
+
+
 ## [2.0.0-alpha.2](https://github.com/empathyco/x/compare/@empathyco/x-deep-merge@2.0.0-alpha.1...@empathyco/x-deep-merge@2.0.0-alpha.2) (2023-06-27)
 
 **Note:** Version bump only for package @empathyco/x-deep-merge